LBOT_MAIN/requirements.txt
# Discord
discord.py>=2.3.0
python-dotenv>=1.0.0

# Database
aiosqlite>=0.19.0

# Image processing (for welcome cards, rank cards)
Pillow>=10.0.0
aiohttp>=3.9.0

# Utilities
python-dateutil>=2.8.0
humanize>=4.9.0

# Optional: For YouTube/Twitch notifications
feedparser>=6.0.0
